Move changes to MapMicro and add three address markers
